api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[asyncio]==5.2.1
aio-pika==9.4.3
aiokafka==0.11.0
strawberry-graphql[fastapi]==0.243.1
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-httpx==0.35.0
